move github utils
